package refactoring: community
