--- src/main/java/com/example/week06/Dto/PostUpdateRequestDto.java
@@ -1,4 +1,0 @@
-package com.example.week06.dto;
-
-public class PostUpdateRequestDto {
-}
