+++ src/main/java/com/example/week06/domain/community/dto/CommentResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.week06.dto;
+package com.example.week06.domain.community.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
