+++ src/main/java/com/example/week06/domain/community/dto/PostUpdateRequestDto.java
@@ -1,0 +1,4 @@
+package com.example.week06.domain.community.dto;
+
+public class PostUpdateRequestDto {
+}
