+++ src/main/java/com/example/week06/domain/community/dto/ResponseMessageDto.java
@@ -1,4 +1,4 @@
-package com.example.week06.dto;
+package com.example.week06.domain.community.dto;
 
 import lombok.Getter;
 import lombok.Setter;
