+++ src/main/java/com/example/week06/domain/community/repository/CommentRepository.java
@@ -1,8 +1,8 @@
-package com.example.week06.repository;
+package com.example.week06.domain.community.repository;
 
-import com.example.week06.model.Comment;
-import com.example.week06.model.Post;
-import com.example.week06.model.User;
+import com.example.week06.domain.community.entity.Comment;
+import com.example.week06.domain.community.entity.Post;
+import com.example.week06.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
