+++ src/main/java/com/example/week06/domain/community/service/PostService.java
@@ -1,17 +1,18 @@
-package com.example.week06.service;
+package com.example.week06.domain.community.service;
 
-import com.example.week06.dto.PostDetailsResponseDto;
-import com.example.week06.dto.PostRequestDto;
-import com.example.week06.dto.PostResponseDto;
-import com.example.week06.dto.ResponseMessageDto;
-import com.example.week06.model.Comment;
-import com.example.week06.model.Post;
-import com.example.week06.model.User;
-import com.example.week06.repository.AttachmentRepository;
-import com.example.week06.repository.CommentRepository;
-import com.example.week06.repository.PostRepository;
-import com.example.week06.repository.UserRepository;
-import com.example.week06.security.UserDetailsImpl;
+import com.example.week06.domain.community.dto.PostDetailsResponseDto;
+import com.example.week06.domain.community.dto.PostRequestDto;
+import com.example.week06.domain.community.dto.PostResponseDto;
+import com.example.week06.domain.community.dto.ResponseMessageDto;
+import com.example.week06.domain.community.entity.Comment;
+import com.example.week06.domain.community.entity.Post;
+import com.example.week06.domain.user.entity.User;
+import com.example.week06.global.AttachmentRepository;
+import com.example.week06.domain.community.repository.CommentRepository;
+import com.example.week06.domain.community.repository.PostRepository;
+import com.example.week06.domain.user.repository.UserRepository;
+import com.example.week06.global.AttachmentService;
+import com.example.week06.global.security.UserDetailsImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
